Split the lz compression tools out of gfx.py.
